split file, comment, cleanup, small restyle
