+++ modules/core/client/less/layout/header.less
@@ -112,7 +112,7 @@
 
     .dropdown-menu {
       margin-top: 5px;
-      margin-right: 10px;
+      margin-right: 5px;
       min-width: 100%; // We'd like dropdown-menu to be similar size to button. But it's 10px wider.
     }
   }
